Rename overlay core module
