+++ scripts/core/overlay_core.py
